fptu-test: fix `<<` minor UB.

+++ externals/libfptu/test/0trivia.cxx
@@ -265,7 +265,7 @@ TEST(Trivia, time_ns2fractional) {
     for (int offset_42 = -42; offset_42 <= 42; ++offset_42) {
       SCOPED_TRACE("base_2log " + std::to_string(base_2log) + ", offset_42 " +
                    std::to_string(offset_42));
-      const uint64_t ns = (1 << base_2log) + offset_42;
+      const uint64_t ns = (uint64_t(1) << base_2log) + offset_42;
       if (ns >= 1000000000)
         continue;
       SCOPED_TRACE("ns " + std::to_string(ns) + ", factional " +
@@ -282,7 +282,8 @@ TEST(Trivia, time_fractional2ns) {
     for (int offset_42 = -42; offset_42 <= 42; ++offset_42) {
       SCOPED_TRACE("base_2log " + std::to_string(base_2log) + ", offset_42 " +
                    std::to_string(offset_42));
-      const uint64_t fractional = uint32_t((1 << base_2log) + offset_42);
+      const uint64_t fractional =
+          uint32_t((uint64_t(1) << base_2log) + offset_42);
       SCOPED_TRACE("fractional " + std::to_string(fractional) + ", ns " +
                    std::to_string(fractional * scale));
       const uint64_t probe = floor(fractional * scale);
@@ -297,7 +298,7 @@ TEST(Trivia, time_us2fractional) {
     for (int offset_42 = -42; offset_42 <= 42; ++offset_42) {
       SCOPED_TRACE("base_2log " + std::to_string(base_2log) + ", offset_42 " +
                    std::to_string(offset_42));
-      const uint64_t us = (1 << base_2log) + offset_42;
+      const uint64_t us = (uint64_t(1) << base_2log) + offset_42;
       if (us >= 1000000)
         continue;
       SCOPED_TRACE("us " + std::to_string(us) + ", factional " +
@@ -314,7 +315,8 @@ TEST(Trivia, time_fractional2us) {
     for (int offset_42 = -42; offset_42 <= 42; ++offset_42) {
       SCOPED_TRACE("base_2log " + std::to_string(base_2log) + ", offset_42 " +
                    std::to_string(offset_42));
-      const uint64_t fractional = uint32_t((1 << base_2log) + offset_42);
+      const uint64_t fractional =
+          uint32_t((uint64_t(1) << base_2log) + offset_42);
       SCOPED_TRACE("fractional " + std::to_string(fractional) + ", us " +
                    std::to_string(fractional * scale));
       const uint64_t probe = floor(fractional * scale);
@@ -329,7 +331,7 @@ TEST(Trivia, time_ms2fractional) {
     for (int offset_42 = -42; offset_42 <= 42; ++offset_42) {
       SCOPED_TRACE("base_2log " + std::to_string(base_2log) + ", offset_42 " +
                    std::to_string(offset_42));
-      const uint64_t ms = (1 << base_2log) + offset_42;
+      const uint64_t ms = (uint64_t(1) << base_2log) + offset_42;
       if (ms >= 1000)
         continue;
       SCOPED_TRACE("ms " + std::to_string(ms) + ", factional " +
@@ -346,7 +348,8 @@ TEST(Trivia, time_fractional2ms) {
     for (int offset_42 = -42; offset_42 <= 42; ++offset_42) {
       SCOPED_TRACE("base_2log " + std::to_string(base_2log) + ", offset_42 " +
                    std::to_string(offset_42));
-      const uint64_t fractional = uint32_t((1 << base_2log) + offset_42);
+      const uint64_t fractional =
+          uint32_t((uint64_t(1) << base_2log) + offset_42);
       SCOPED_TRACE("fractional " + std::to_string(fractional) + ", ms " +
                    std::to_string(fractional * scale));
       const uint64_t probe = floor(fractional * scale);
